REFACTOR : category 인식 못 하는 버그 해결
